+++ lib/db/src/schema/index.ts
@@ -42,6 +42,7 @@ export * from "./workflow";
 export * from "./wagenpark";
 export * from "./organisatie";
 export * from "./werkbak";
+export * from "./uitrol";
 export * from "./werkvoorbereiding";
 export * from "./leveranciers";
 export * from "./artikelen";